+++ control/migrations/000001_create_workers_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS workers (
     id TEXT PRIMARY KEY,
     hostname TEXT NOT NULL,
     ip_address TEXT NOT NULL,
-    cpu_cores INTEGER NOT NULL,
+    cpu_cores FLOAT NOT NULL,
     ram_mb INTEGER NOT NULL,
     token TEXT UNIQUE NOT NULL,
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
